Improved handling of MySQL Server Deployment scale down
When a scale down is requested on the MySQL Server Deployment, the
deployment is now scaled down first before applying any config changes
to the Management node and rest of the cluster. This is to ensure that
the MySQL Servers get shut down before the API sections in the config
are possibly reduced.

Change-Id: I07c8f31f1082dab1541e1b2cd49b703f0bb7fa6e