MySQL Server root password generation
A random alphanumeric password is now generated and stored as a secret
in the k8s cluster. The secret is now mounted as a volume into the MySQL
Server containers and the password itself is projected at /etc/auth
location inside the pods. The next step is to set this password to the
root user during the MySQL Server data directory initialization.

Change-Id: I8d37be642ea9967cfff9d6cf7a7d6be9ca658134