Added a new argument 'scripts_dir' to Ndb Operator
MySQL Server requires couple of scripts to be mounted into the pods for
ndbcluster specific initialisations and other stuffs. For this, the
operator needs to have access to these scripts so that it can mount them
as volumes inside the pods using configmaps. When the operator is run
inside the k8s cluster, from the operator container image, the scripts
will be a part of the image and will accesible from a predefined
location. But when the operator is run outside the k8s cluster, the
location of the directory that has the scripts, has to be passed
manually. The new argument fulfills that purpose.

Also added a .dockerignore file to exclude everything except the files
and folders required for the docker build.