Added ndbcluster specific mysqld initialisation script
The entrypoint script that comes as a part of the standard MySQL
container image does the data directory initialisation and a few other
updates to the users and database. Although that is sufficient for a
vannilla MySQL Server, it is not sufficient or causes issues to a
ndbcluster enabled MySQL Server. Creating a root user at a requested
host for example, is executed by all connected MySQL Servers in
parallel and when they attempt to GRANT ALL privileges to the user,
it causes a conflict as the new NDB_STORED_GRANT is also applied to
the user and all the Servers attempt to distribute the user in parallel.
One other thing is the health check mechanism for ndbcluster enabled
MySQL Server - user defined functions and procedures are required to
write better liveness/readiness probes.

So to improve any such shortcomings, an additional initialisation script
is introduced in this patch. It gets mounted into the
/docker-entrypoint-initdb.d/ of the MySQL Server pods so that the
entrypoint script would execute it as a part of its execution.

This script does the following for now :
 - Creates a new root user at requested host in a safe manner - only one
   MySQL Server creates it and any other Servers connected to the NDB
   cluster gets the changes via distribution. The script achieves it
   through few procedures and functions defined in the script.
 - Creates a new stored function 'mysql.IsNdbclusterSetupComplete()',
   that returns if the ndbcluster specific setup is complete in the
   MySQL Server.

Change-Id: I5805abbef5b5ebcf6a497c53166d68cab4e91db3