Added support to generate a yaml install artifact
A single yaml file that installs all the ndb operator resources in the
default namespace will be generated at the location
artifacts/install/ndb-operator.yaml.

The generation is hooked up with manifests and build targets in Makefile
and are run when there is a change in one of the files affecting the Ndb
resource definitions.

Change-Id: I1ab9caebfcfbcd3cfc3d8d538b9c3bfbfe988142