Improved the timing of Ndb resource status update
When a change is applied to the Ndb resource by the user, the first
reconciliation loop just patches the config map and does the rest of the
other changes over subsequent calls. But the status of the Ndb resource
gets updated immediately after the config map is patched. This patch
fixes this issue by updating the status of the Ndb resource only after
all the changes have been applied to the MySQL Cluster. Patch also
improves couple of comments to the CRD.

Change-Id: I20f230b914ce9282c0004b0681fd1fdd605e2b2c