Improved Events sent by Ndb Operator
When a new spec change is applied to the ndb resource, the first loop
that handles the change, only syncs the config map with the new spec.
The other resources won't be synced yet. Despite that, the first loop
records a 'Synced' event before returning. This might be misleading.

Updated the code to record a Synced event only after actually syncing
all the resources. Also, added a new variant of the Synced event that is
recorded when the Ndb resource is already in sync with the spec.

Also removed an unused method.

Change-Id: Iabb8bc892ccb19f54ec5deed6f47e95787cab7a3