Jira#4658: Fixed a bug in MySQL Server scale down
When the MySQL Server is scaled down, the config in the Mgmd gets
updated and the number of API slots in management config gets reduced.
Although the MySQL deployment is scaled down before this change, there
is a chance that the node ids used by the MySQL Servers, that remain
after the scale down, get removed during this config update. This is
also because MySQL Servers are a deployment and we don't know which
Servers (and thus node ids) will be shut down during a scale down. When
such a node id mismatch occurs, the Ndb cluster might never become fully
ready as there will be issues during the data node restart. Fixed that
by updating the config generation logic to not reduce the number of API
slots during a scale down. Also added a new testcase for testing the
scale up and down of MySQL Servers.

Change-Id: Ib73a30cee5bdbe9674d33f8a621065c68c8d7a62