Updates to e2e tests
To connect to a service, the e2e tests used the load balancer ingress
points available in the service status. This becomes an issue if the
tests are run inside the K8s cluster which doesn't provide a load
balancer. This patch updates the testcases to connect to the Services
via ClusterIPs for such cases.

Change-Id: I3c5d74c4b7620e071b015aff64b1b2590642cca8