Improved the MySQL Cluster health check logic
The controller relies on the ClusterStatus's NumberNodegroupsFullyUp()
method and the SyncContext's checkClusterState() to decide if the
MySQL Cluster running in the K8s is healthy or not. These methods used a
complex logic to detect if a MySQL Cluster is degraded as the NodeGroup
value of a disconnected node was not available to them. But the
NodeGroup of such disconnected nodes are now available in ClusterStatus
after recent improvements to the mgmapi. So, these two methods are now
replaced by a simpler IsHealthy() method by this patch.

Change-Id: I00eda41b43f450fc30b80250c98cbd48cb41fb6f