Updated data node handling during reconciliation
Added a new method GetNodesGroupedByNodegroup() to ClusterStatus to
replace the existing topology logic. The new method returns an array of
array of nodes sorted and grouped based on nodes ids and nodegroup
numbers. The controller now uses this value to decide which data nodes
to pick and validate during a reconciliation. This patch also removes
the topology logic and the related files from mgmapi.

Change-Id: I16372cc6b9e3f46735de0fcdfc04be1e8cccdb30