Changed the CRD name from Ndb to NdbCluster
Renamed the CRD name to NdbCluster to make it more clear what exactly
the resource defines. Lots of internal code and comments still refer to
the resource as an Ndb type. They are not yet an issue as there is only
one Ndb CRD but they will be updated over time.

Changes :
 - Renamed Ndb CRD to NdbCluster
 - Added 'ndb' as a shortname for the NdbCluster resource
 - Renamed NdbSpec to NdbClusterSpec
 - Renamed NdbStatus to NdbClusterStatus

Change-Id: I52995efb543a384511296d2cef69f4c25671e967