Fixed a read issue in Mgmapi sendCommand
The sendCommand uses the net.Conn's Read method to extract the reply
sent by the Management Server. Once all the reply has been read, any
further read would throw a i/o timeout, which basically is a
os.ErrDeadlineExceeded error. The sendCommand wrongly treated this as an
error rather than the end of reply indication it is. Due to this, even
after a successful read, the sendCommand failed on encountering the
timeout error, causing the reconciliation to fail.

Fixed that by improving the handling of os.ErrDeadlineExceeded. Also
removed the retry logic from the sendCommand method as that is not
necessary.

Change-Id: I02f4fd5f0e55f998f70ec22289612f520464029d