MYSQLCLUSTER-4732 MADE MYSQL ROOT HOST CONFIGURABLE FROM NDB CRD
Currently, for mysqld deployment it is not possible for user to
configure host that is to be used for MySQL root account, and a
default value "%" is used as the host name.

In this change, support is added to configure mysqld's root hostname
from Ndbcluster custom resource definition by specifying "rootHost"
parameter (immutable) under mysqld spec.

Change-Id: I78367adfa7b1a6f65035029e3da06c40011b7c06