MYSQLCLUSTER-4869: Operator should support multiple MySQL Clusters
When multiple NdbCluster objects are created, the NDB Operator sets up a
MySQL Cluster for the first NdbCluster object. But for the other
NdbCluster objects, it throws an error saying the resource already
exists. This was caused due to the name of the first NdbCluster
incorrectly being stored in the MySQLServerDeployment and the
baseStatefulSet resources. The controller rather than using the name of
the current NdbCluster being synced, used the incorrect name to find out
if a K8s resource exists.

This patch fixes that by removing the incorrect names stored in the
MySQLServerDeployment and baseStatefulSets and making the controller to
rely on the name of the current NdbCluster being handled. Also added a
testcase.

Change-Id: Ic3e41e28e21ca0110f52310db3740c8c0d2302bc