Improve interrupt handling in e2e tests tool
Added a new entrypoint script to the e2e-tests docker image to properly
handle the interrupt signal and abort the Ginkgo test. Also updated the
e2e tests tool to delete the test pod when a user aborts the test. These
changes will now enable a proper cleanup of resources when a test is
aborted midway.

Change-Id: I4edcbbd47b5690bd7a105463074f55fdc04caef2