MYSQLCLUSTER-5402: e2e test should fail when the operator crashes
If the NDB Operator exited due to an error, it went unnoticed by the e2e
framework. As the operator is running as a deployment, in certain cases,
the crashed operator was able to restart and continue its work without
further errors. This allowed the testcases to succeed despite the NDB
Operator throwing up an error and crashing midway.

This patch adds a new go routine that watches the NDB Operator pod and
fails the testcase if the pod exits due to an error. This patch also
adds another go routine that collects the log of the NDB Operator. The
logs will be printed to the ginkgo log if the testcase fails.

Change-Id: I0eddccfd07a9e5282b7119ad6402d1e315e332f4