SyncSuccess event sent twice if the cache object is outdated
The events are sent based on any changes detected between the original
NdbCluster object's status retrieved from the go client cache and a
copy of it whose status was updated by the SyncContext. There is a
chance that the cache object is outdated causing a conflict during the
status update. The status update is then retried with the latest
version of the NdbCluster object retrieved from the K8s Server but
without checking again if there has been an actual change in status
field values. This caused the same status change to be updated twice in
the K8s Server and the SyncSuccess event to be recorded twice.

This patch fixes that by comparing the statuses of the two NdbCluster
objects again after a conflict before sending the update to the K8s
Server. The event sending logic now takes this into account as well and
records the SyncSuccess event only once when the actual sync completes.

Change-Id: I94fb10d7a03c0f16e1d1335067451bbb7d550d55