MYSQLCLUSTER-5150 Better handling of NdbCluster spec sync
Part I : MySQL Cluster config should not be updated if only the MySQL
Server spec has changed.

The MySQL Cluster config version was previously dependent on the
NdbCluster Generation number and this triggered a rolling restart of all
management and data nodes even for a MySQL Server config change in
NdbCluster spec. This patch makes that config version independent of the
Generation Number. It is updated only when there is an actual change to
the MySQL Cluster config and thus triggering a rolling restart only when
a config change has to be applied to MySQL Cluster.

Change-Id: Iaa2036064a16fc36c33b344ecc897d87cdfbf63f