MYSQLCLUSTER-5150 Better handling of NdbCluster spec sync
Part II : MySQL Server deployment template should be updated only when
there is a change to the MySQL Server configuration

The MySQL Server deployment spec template gets updated even for a MySQL
Server scale-up request and the servers get restarted. This commit
updates the deployment patching logic to update the template spec only
when there is either a MySQL Cluster config change or a MySQL Server
config change. So the deployed servers get restarted only if there is a
config change that affects them. A MySQL Server scale-up or a scale-down
request will be handled without restarting the existing nodes if there
are enough free api slots in the MySQL Cluster config.

Change-Id: Ic1f872885bd5ec5af3614bc8cd50219b9397d3a9