Docker entrypoint script should fail when the e2e tests fail
The docker entrypoint succeeds even if the ginkgo tests run by it
failed. The wait call used by the docker entrypoint exits with zero exit
status as the pid is not passed to it, causing the pod running the tests
to succeed. Fixed that by properly passing the ginkgo test pid to the
wait command.

Change-Id: Id9a1f68f128aa71f838c82b2db157cd19e38f6de