MYSQLCLUSTER-4875 MySQL root host and password should be mutable [part 1]
Problem
=======
NdbCluster's spec.mysqld.rootHost field is immutable and any attempt
to change it after the initial start up will be rejected by the K8s
Admission Controller. The root user is created by an init script when
the MySQL Server pod is started and currently there is no way to change
it from the Ndb Operator. This patch adds support to make the field
mutable and updates the Ndb Operator to handle changes to the root
host even after an initial start up.

Fix
===
The init script now creates a master user - 'ndb-operator-user', which
is used by the operator to create, update and delete the root user with
the root host specified in the NdbCluster spec. This patch also removes
the workaround introduced for Bug#32424653 in the MySQL Server init
script as that is fixed now.

Change-Id: I422d39192a952750a9fde62bef6ace0bfcdf7cfb