MYSQLCLUSTER-5699 Data node pods reconnect too early after a pod deletion
If a data node pod gets deleted, the new pod started by the StatefulSet
controller to take its place, should wait for the other existing data
nodes to complete handling the previous node failure. The new pod should
also wait until the internal K8s DNS gets updated with its hostname and
recent IP address. Failing to wait for either of these will cause the
data node to fail with an error.

This patch adds an init container to the Data Node StatefulSets that
uses a bash script to wait until the above conditions are satisfied. The
actual data node container will start only after the init container
succeeds. This will ensure that the data node reconnects only when other
parts of the system are ready for it.

This patch also renames datanode-healthcheck.sh script to
datanode-startup-probe.sh.

Change-Id: I8c78e1a839372e71f06d636a667ad2d33f89f319