MYSQLCLUSTER-5660 Handle MySQL Cluster nodes' pod spec update
Any changes to the management and data nodes statefulsets are currently
not possible as there is no support for patching them in the NDB
Operator. Any updates to the MySQL Cluster config was resolved simply by
stopping the nodes by issuing an mgm command and letting the statefulset
controller restart the nodes with the latest config. This does not work
for statefulset definition updates as the node has to be explicitly
deleted to allow the statefulset controller restart them with the new
definition. This patch adds support for such a statefulset update
mechanism and replaces the exisiting mechanism with it.

Changes :
 - Added support for patching a statefulset if the dependent CRD spec
   changes.
 - Management node statefulset uses the default RollingUpdate strategy
   and any update is automatically rolled out by the statefulset
   controller.
 - Data node statefulset uses the legacy OnDelete strategy and the
   update is handled by the Ndb Operator. The pods are updated based on
   the statefulset revision hash. The pods with the older version are
   deleted by the operator and the controller restarts them with newer
   versions. This is done without affecting the availability of the
   MySQL Cluster. One data node per nodegroup is picked up per loop and
   are updated together.
 - The MySQL Cluster config version is now annotated to the statefulset
   pod definition so that any changes to the config will also trigger an
   update to the statefulset definition.

Change-Id: Ie35bb58ceefa67ffdc903a26aa760050fcbbd9ac