MYSQLCLUSTER-5742 Use ndbmtd in data node pod names instead of ndbd
The data node statefulsets and their pods have 'ndbd' in their generated
names despite using ndbmtd to run the data nodes. This might lead to
confusion as in which executable is used to start the data nodes. So to
make it more clear, this patch updates the generated names of the
statefulset and their pod names to use 'ndbmtd' instead of 'ndbd'.

The actual change is in pkg/constants/constants.go file where the
NdbNodeTypeNdbmtd is now defined as "ndbmtd". This patch also makes
other similar updates to code, comments, and docs, wherever possible.

Change-Id: Ic70460a7dfbb70f7888d95efc7ccac96bc1331c0