Fixed a logic bug in ensureWorkloadsReadiness
The ensureWorkloadsReadiness method checked only if the management
statefulset was ready and did not check if any update is being rolled
out to the statefulset. This sometimes caused the operator to
prematurely continue with the reconciliation causing errors. This patch
updates the method to properly check for any ongoing management
statefulset updates in progress.

Change-Id: I94d05e4c0028db1260f27f9e139a6de6dfb3371b