MYSQLCLUSTER-4871 Replace default MySQL docker entrypoint script
Added a new bash script to initialize the data directory of the MySQL
Servers. This script is now used by a new init container in the MySQL
StatefulSet to initialise the data directory. Since the init container
initialises the data directory, the main container starts the MySQL
Server by using the executable directly rather using the entrypoint
script that comes with the image.

Other changes :
 - The ndbcluster-init-script is now removed as its contents are now
   merged into the new bash script.
 - The MySQL Root password is projected via environment variable rather
   than a file.
 - Updated the MySQL healthprobe to use the correct location of the init
   file and config.
 - Update the names of helper scripts to reflect the executable on which
   they work.

Change-Id: Id41684959c378cd7667fbda68c570fb72a500c74