Increase the timeout for the Mgmd startup probe script
Mgmd startup probe uses "getent host" to resolve the hostname of the
mgmd. In some slower machines this command takes more than 8 sec to
return. But the default timeout for the entire Mgmd startup probe
script is 3 sec. So, in slower machines Mgmd startup probe will
always fail.

Fixed the issue by increasing the timeout value of the Mgmd startup
script to 10 sec.

Change-Id: Ib296a67e559f308a11d64aeec837b760b490d7ae