MYSQLCLUSTER-6241 TEST WITH LOCAL DOCKER REPOSITORY
Change all the reference to public Docker Hub images in the code.
  - use OCR for all possible images
  - for the ones that does not have OCR image, push the image to
OCIR(internal) and use OCIR.
  - mask the internal image reference in deliverables. Update the
makefile to capture the env variables into docker build argument.
So, when the env variable is set with an internal iamge name, the
internal image will be used for the docker build else the image
specified in the Dockerfile will be used.
  - record CRD docs file

Change-Id: I3ac522282f7bdecdea480490ee5de794976515e0