MYSQLCLUSTER-6644 Fix CA certificates in Webhook server
Problem:
=======
  When the ndb-operator is installed using OLM, it cannot handle new
ndbcluster CRD (Custom Resource Definition) requests from the user.

Description:
===========
  During startup, the HTTP server inside the ndb-operator generates TLS
certificate and key and updates all the webhook configurations by
adding the created certificate to them. This means that the webhook
configurations use these certificates when creating HTTP requests, and
the same certificates are used on the server side for validation.

  When installing the ndb-operator using the OLM (Operator Lifecycle
Manager), the ownership of the webhook configurations lies with the
CSV (ClusterServiceVersion). As a result, any modifications made to the
webhook configurations will not be reflected because the CSV controls
the Kubernetes objects created by it. Consequently, the authentication
step fails for the HTTP requests sent by the webhook configurations.

Fix:
===
  OLM creates and mount a self-signed Certificate Authority (CA) to
each deployment that contains a webhook. It mounts the TLS certificates
and keys to a hardcoded path in the pods. The ndb-operator will first
check this path for certificate. If it finds the certificate, it uses
them for validation. If it doesn't find the certificate, it creates new
certificates and updates all the webhook configurations. This ensures
that the ndb-operator can work with all modes of installation.

Change-Id: I9c5aeb0ca52b33926621c72639a0300b925a8c15