MYSQLCLUSTER-6912 Mysql pod does not start with the mysql cluster 8.1.0 container image
The Ndb-operator employs an initialization script containing multiple
MySQL statements, which is executed as an init container in the mysqld
pod.

However, a bug is encountered in MySQL Server 8.1.0. When a space is
present before the "#" character in a comment preceding the "DELIMITER"
command, the "DELIMITER" command is not executed as intended. Instead,
the command is skipped, leading to discrepancies when executing
subsequent statements.

As a result, the initialization script executed in the init container
becomes faulty, preventing the mysqld pod from starting up successfully.

To address the issue, the bug was resolved by removing the comment at
the beginning of the MySQL statement, allowing the "DELIMITER" command
to function correctly. This ensures proper execution of the
initialization script and resolves the problem with the mysqld pod
startup.

Change-Id: I3bdf9f8b0721e4fc9f2deefd5621e9e330ec4742