MYSQLCLUSTER-6983 Fix image pull issue in ndb-cluster pods
When deploying the ndb-operator using an image secret, any attempt to
create the ndbcluster CRD in a multi-node Kubernetes cluster will result
in pod initialization failures due to image pull issues. The root cause
of this problem lies in the ndb-operator code, which neglects to include
the necessary image pull secret in the pod specifications. The missing
image pull secret prevents the successful retrieval of the ndb-operator
image, crucial for the initialization container within the pods. As a
consequence, when pods are created on nodes other than the one hosting
the ndb-operator, the required image pull secret is absent, resulting in
image retrieval failures.

This issue is now fixed by updating the StatefulSet to incorporate the
necessary image pull secret for the ndb-operator image.

Change-Id: I970d1a28f7abf2b5512ca34da9143932e693ae6e