MYSQLCLUSTER-7007 Fix mysqld initscript logic
The Ndb-Operator currently triggers the launch of the mysqld container
even in cases where errors occur during the execution of user-provided
init scripts.

The current patch addresses this issue by rectifying the error handling
logic within the init scripts. With the updated implementation, the
error handling logic effectively halts the execution of the init script
upon encountering any failure. As a result, the init container is
appropriately halted, ensuring that the mysqld container is not launched
when issues are detected within the init scripts.

Change-Id: Iabeffec3142b74d50a7273af2458eb7a52502a29