MYSQLCLUSTER-9057 Harden operator security contexts
Hardens the security context of the NDB Operator and adds extends it, as
an opt-in, to the deployed NDB Cluster pods, following the
least-privilege principle. This should allow NDB deployments to comply
with more strict security policies.

The main changes are:
- Run pods as non-root user. This requires that NDB Cluster processes
  run with a UID and GID different from 0. By default, NDB Cluster
  processes will run with UID and GID equals to 27. However, this can be
  configured, and there is an option to allow the platform to
  automatically assign a value, when supported;
- Disallow privilege escalation;
- Drop all extra Linux capabilities;
- Restrict the Secure Computing Mode (seccomp) to runtime default;
- Set the root file system as read-only. The mysqld containers are an
  exception because the implementation requires writing to some
  locations outside the volumes. It should be possible to use local
  volumes to remove this writing requirement, but is should be handled
  as another task.

The hardened security context is enabled by default when deploying the
NDB Operator pods but disabled for the NDB Cluster pods. This was done
to avoid breaking any existing systems after upgrading, specially
because the issues can happen when NDB Cluster pods restart. This may be
long after the restart, thus making the debugging harder. The possible
problem is related with the ownership of the files and directories
accessed by the NDB Cluster processes when using Persistent Volumes.
This may require some changes to the ownership and/or permissions in the
existing files and directories to allow the access by the NDB Cluster
processes running with a specific UID and GID.

An additional change was required in the mysqld init container so that
the 'mysql-init-script.sh' could be executed by an user other than root.
This file is inside a configuration map which is mounted in the
container and owned by the root user which is the only one with
permissions to execute. Several experiments have shown that the best
solution was to also give execution permissions to the group.

Change-Id: I96160b832f46dda1352db9ba4d8907fb68c90cec